+++ core-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao core bundle change log
 
+### DEV
+
+ * Correctly assign the CSS media type in the combiner.
+
 ### 4.4.2 (2017-07-25)
 
  * Adjust the command scheduler listener so it does not rely on request parameters (see #955).